Add files via upload
UPDATED VERSION: Added comments  to the files I have been working on, added window.location.reload() to the login page and added a div that had disappeared but is necessary in order to hide some buttons from users that are not admin. Also added class = "translatable" to the language buttons as it had disappeared somehow.